% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ak1km}
\alias{ak1km}
\title{Alaska 1-km ALFRESCO template raster layer.}
\format{A RasterLayer.}
\usage{
ak1km
}
\description{
A raster layer of the 1-km Alaska classic "statewide" ALFRESCO model spatial domain.
All non-NA values are set to 1.
}
\keyword{datasets}
