% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{akcan1km}
\alias{akcan1km}
\title{Alaska/western Canada 1-km ALFRESCO template raster layer.}
\format{A RasterLayer.}
\usage{
akcan1km
}
\description{
A raster layer of the 1-km Alaska/western Canada PRISM extent used for the SNAP climate downscaling domain, resampled to the 1-km ALFRESCO model spatial domain.
All non-NA values are set to 1.
}
\keyword{datasets}
