% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{akcan2km}
\alias{akcan2km}
\title{Alaska/western Canada 2-km downscaled climate template raster layer.}
\format{A RasterLayer.}
\usage{
akcan2km
}
\description{
A raster layer of the 2-km Alaska/western Canada PRISM extent used for the SNAP climate downscaling spatial domain.
All non-NA values are set to 1.
}
\keyword{datasets}
