% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{snapgrid}
\alias{snapgrid}
\alias{snapgrid-package}
\title{snapgrid: Common SNAP geographic raster layer objects.}
\description{
The snapgrid package contains a collection of rasterized maps focused on
Alaska and Canada, including vegetation input for the ALFRESCO wildfire model,
fire management options map layers, domain template layers for the Alaska "statewide" classic ALFRESCO domain,
1-km ALFRESCO and 2-km climate templates and more.
}
\details{
snapgrid is a member package in the data sector of the SNAPverse.
All maps are in the NAD83 Alaska Albers Equal Area Conic projection for convenience and conformity,
but cover multiple extents and users should be familiar with the layers already in order to ensure proper usage context.

There following maps are available:
\describe{
\item{\code{akcan2km}}{Alaska/western Canada 2-km downscaled climate domain template layer.}
\item{\code{akcan1km}}{Alaska/western Canada 1-km ALFRESCO domain template layer.}
\item{\code{ak1km}}{"Statewide" classic ALFRESCO domain template layer.}
\item{\code{swveg}}{Vegation map layer for the classic "statewide" ALFRESCO model spatial domain.}
\item{\code{swfmo}}{Statewide fire management options zones map layer.}
\item{\code{swratios}}{Statewide fire management options zone ratios map layer.}
\item{\code{swflam}}{Statewide GBM vegetation-mediated, climate-driven landscape flammability mask map layer.}
}
}
