% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{snapgrids}
\alias{snapgrids}
\title{Basic metadata for all data sets in snapgrid}
\usage{
snapgrids()
}
\value{
a data frame.
}
\description{
This function returns a data frame with basic meta data for all data sets in \code{snapgrid}.
This includes data object names, spatial domain, spatial resolution, and short description.
}
\examples{
snapgrids()
}
