% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swflam}
\alias{swflam}
\title{GBM vegetation-mediated climate-driven landscape flammability.}
\format{A RasterLayer.}
\usage{
swflam
}
\description{
A raster layer of the ALFRESCO gbm (gradient boosting modeled) vegetation-mediated climate-driven landscape flammability.
This map describes the spatial domain of the flammable region for the statewide ALFRESCO runs for which this layer is used.
It is based on the \code{swveg} data set, excluding the Alaska range area.
}
\keyword{datasets}
