% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swfmo}
\alias{swfmo}
\title{Alaska Statewide ALFRESCO fire management options zones.}
\format{A RasterLayer.}
\usage{
swfmo
}
\description{
A raster layer of the ALFRESCO fire management option zones,
conforming to the classic "statewide" spatial domain.
The ID codes for the raster layer are as follows:
}
\details{
\describe{
\item{\code{0}}{No fire management}
\item{\code{1}}{Limited}
\item{\code{2}}{Modified}
\item{\code{3}}{Critical}
\item{\code{4}}{Full}
\item{\code{4}}{Other}
}
}
\keyword{datasets}
