% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swfmoBuffer}
\alias{swfmoBuffer}
\title{Alaska Statewide ALFRESCO fire management options buffered zone.}
\format{A RasterLayer.}
\usage{
swfmoBuffer
}
\description{
A raster layer of the ALFRESCO fire management options using 15-km buffered management areas,
conforming to the classic "statewide" spatial domain.
The ID codes for the raster layer are as follows:
}
\details{
\describe{
\item{\code{0}}{No fire management}
\item{\code{1}}{Limited}
\item{\code{2}}{Modified}
\item{\code{3}}{Critical}
\item{\code{4}}{Full}
}
}
\keyword{datasets}
