% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swratios}
\alias{swratios}
\title{Alaska Statewide ALFRESCO fire management options zone ratios.}
\format{A RasterLayer.}
\usage{
swratios
}
\description{
A raster layer of the ALFRESCO fire management options zone ratios conforming to the classic "statewide" spatial domain.
All values are greater than or equal to one. Grid cells equal to one do not experience fire suppression in ALFRESCO.
ALl other cells experience fire suppression to a degree based on their relative ratios.
The mechanism for this in ALFRESCO is that this map layer is used to weight the fire sensitivity and/or ignition factor ALFRESCO
input geotiffs.
}
\details{
Ratios for full and critical zones are 1.5 and 1.75, respectively. All other areas are set to 1.0.
}
\keyword{datasets}
