% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swveg}
\alias{swveg}
\title{Alaska Statewide ALFRESCO vegetation input.}
\format{A RasterLayer.}
\usage{
swveg
}
\description{
A raster layer of the ALFRESCO vegetation input conforming to the classic "statewide" spatial domain.
The ID codes for the raster layer are as follows:
}
\details{
\describe{
\item{\code{0}}{No vegetation/non-burnable area, e.g. mountain or water body}
\item{\code{1}}{Alpine tundra}
\item{\code{2}}{Black spruce}
\item{\code{3}}{White spruce}
\item{\code{4}}{Deciduous tree species}
\item{\code{5}}{Shrub tundra}
\item{\code{6}}{Graminoid tundra}
\item{\code{7}}{Wetland tundra}
}
}
\keyword{datasets}
