% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swfmoBuffer}
\alias{swfmoBuffer}
\title{Alaska Statewide ALFRESCO fire management options buffered zone.}
\format{A RasterLayer.}
\usage{
swfmoBuffer
}
\description{
A raster layer of the ALFRESCO fire management options 15-km buffered zone for the union of modified, full and critical management areas,
conforming to the classic "statewide" spatial domain. All non-NA values are 1.
}
\keyword{datasets}
